tests: Move core dumps to a separate crate
Keep main repo size down post-squash.
