+++ Cargo.toml
@@ -31,9 +31,8 @@ libarchive = "0.1.1"
 libarchive3-sys = "0.1.2"
 
 [dev-dependencies]
-flate2 = "0.2.14"
 byteorder = "0.5"
-lazy_static = "0.2.1"
+rbspy-testdata = { git = "https://github.com/rbspy/rbspy-testdata.git" }
 tempdir = "0.3.4"
 
 [build-dependencies]
